pyproject.toml
[tool.poetry]
name = "short-circuit-me"
version = "0.3.3"
description = "short-circuit-me poetry"
authors = ["John Dutchover <johndutchover@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "scm-poetry"}]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11.0"
fastapi = "^0.97.0"
slack-bolt = "^1.18.0"
pip-tools = "^6.13.0"
python-dotenv = "^1.0.0"
pydantic = ">=1.8.0,<2.0.0"
uvicorn = ">=0.15.0,<0.16.0"
requests = "^2.3.0"
streamlit = "^1.23.1"
watchdog = "^3.0.0"
asyncio = "^3.4.3"
numpy = "^1.24.3"
pandas = "^2.0.2"
plotly = "^5.14.1"
bokeh = "2.4.3"
motor = "^3.1.2"
pymongo = "^4.3.3"
streamlit-extras = "^0.2.7"
load-dotenv = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
typer = "^0.9.0"
flake8 = "^6.0.0"
black = "^23.3.0"
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
